+++ skbio/util/_misc.py
@@ -12,6 +12,8 @@ import hashlib
 from os import remove, makedirs
 from os.path import exists, isdir
 from functools import partial
+from warnings import warn
+
 
 
 def cardinal_to_ordinal(n):
@@ -268,6 +270,17 @@ def find_duplicates(iterable):
 def flatten(items):
     """Removes one level of nesting from items
 
+    .. note:: Deprecated in scikit-bio 0.2.3-dev
+       ``util.flatten`` will be removed in scikit-bio 0.3.1
+       it is being deprecated in favor of solutions present
+       in the standard python library.
+       Please refer to the following links for good alternatives:
+       http://stackoverflow.com/questions/952914/making-a-flat-list
+       -out-of-list-of-lists-in-python
+       http://stackoverflow.com/questions/406121/flattening-a-shallow-
+       list-in-python.
+
+
     Parameters
     ----------
     items : iterable
@@ -287,6 +300,13 @@ def flatten(items):
     ['a', 'b', 'c', 'd', 1, 2, 3, 4, 5, 'x', 'y', 'foo']
 
     """
+    warn("flatten is deprecated. Please refer to the following links for "
+         "solutions from the standard python library "
+         "http://stackoverflow.com/questions/952914/making-a-flat-list"
+         "-out-of-list-of-lists-in-python "
+         "http://stackoverflow.com/questions/406121/flattening-a-shallow-"
+         "list-in-python.", DeprecationWarning)
+
     result = []
     for i in items:
         try: